+++ src/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/13 17:58:25 by mcogne--          #+#    #+#             */
-/*   Updated: 2024/10/13 18:03:08 by mcogne--         ###   ########.fr       */
+/*   Updated: 2024/10/14 15:39:41 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
